% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\docType{data}
\name{bmi}
\alias{bmi}
\title{BMI of 1,781 mothers}
\format{A data frame with 1781 rows and 4 variables:
\describe{
  \item{bmi}{Body Mass Index (15.04-45.18)}
  \item{inv_bmi}{inverse Body Mass Index (0.022-0.066)}
  \item{parity}{ parity of the mother (primi, multi)}
  \item{group_par}{0 = primipari, 1 = multipari}
}}
\usage{
bmi
}
\description{
A dataset containing the Body Mass Index (BMI) of mothers at the beginning of their pregnancy and a variable showing if it is the first pregnancy (parity=0) or a subsequent (parity>0).
}
\keyword{datasets}
