% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi2.R
\docType{data}
\name{bmi2}
\alias{bmi2}
\title{BMI of 1,560 mothers}
\format{A data frame with 1560 rows and 4 variables:
\describe{
  \item{bmi}{Body Mass Index (15.04-45.18)}
  \item{inv_bmi}{inverse Body Mass Index (0.022-0.066)}
  \item{employ}{employment status (unemployed, employed)}
  \item{group_emp}{1 = employed, 2 = unemployed}
}}
\usage{
bmi2
}
\description{
A dataset containing the Body Mass Index (BMI) of employed and unemployed mothers.
}
\keyword{datasets}
