% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwsmoke.R
\docType{data}
\name{bwsmoke}
\alias{bwsmoke}
\title{Birth weight of 1,458 babies}
\format{A data frame with 1458 rows and 3 variables:
\describe{
  \item{birthwt}{birth weight in gram (1780-4870)}S
  \item{smoke}{smoking status of the mother (smoker, non-smoker)}
  \item{gest}{gestational age (37-44.43)}
}}
\usage{
bwsmoke
}
\description{
A dataset containing the smoking status of the mother during their pregnancy, the gestational age and the birth weight of their babies.
}
\keyword{datasets}
